+++ connectToAWSServerViaSSH.sh
@@ -1,1 +1,1 @@
-ssh -i ~/.ssh/linux-demo.pem ec2-user@18.216.94.221
+ssh -i ~/.ssh/linux-demo.pem ec2-user@52.15.176.248