+++ src/main/java/com/github/jeffmagina/takeout/restaurant/customer/Customer.java
@@ -1,4 +1,4 @@
-package com.github.jeffmagina.restaurant.customer;
+package com.github.jeffmagina.takeout.restaurant.customer;
 
 public class Customer {